chore(b3fs): move HashTree to seperate file to separate it from tokio
