+++ lib/b3fs/src/bucket/mod.rs
@@ -255,28 +255,6 @@ pub(crate) mod tests {
         fs::remove_dir_all(&temp_dir);
     }
 
-    #[tokio::test]
-    async fn test_read_verify() {
-        // Put file into blockstore
-        let n_blocks = 10;
-        let mut temp_dir = temp_dir();
-        temp_dir.push("test-read-verify");
-        let bucket = Bucket::open(&temp_dir).await.unwrap();
-        let mut writer = FileWriter::new(&bucket).await.unwrap();
-        let mut data = get_random_file(8192 * n_blocks);
-        let mut number_blocks = n_blocks;
-        writer.write(&data).await.unwrap();
-        let hash = writer.commit().await.unwrap();
-        println!("{hash:?}");
-
-        // Read from the blockstore
-        let load_content = bucket.get(&hash).await.unwrap();
-        let file_read = load_content.into_file().unwrap();
-
-        let blocks = file_read.num_blocks();
-        let tree = file_read.hashtree().await.unwrap();
-    }
-
     pub(crate) fn get_random_file(size: usize) -> Vec<u8> {
         let mut data = Vec::with_capacity(MAX_BLOCK_SIZE_IN_BYTES);
         for _ in 0..size {
